# docs/api/clubspot.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ClubSpot.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/bookings": {
      "post": {
        "tags": [
          "bookings"
        ],
        "operationId": "CreateBooking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingCreatedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
    "/api/portal/{clubSlug}/catalog": {
      "get": {
        "tags": [
          "portal"
        ],
        "operationId": "GetPortalCatalog",
        "parameters": [
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalCatalogResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/{clubSlug}/availability": {
      "get": {
        "tags": [
          "portal"
        ],
        "operationId": "GetPortalAvailability",
        "parameters": [
          {
            "name": "sport",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalAvailability"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/portal/{clubSlug}/bookings/{id}": {
      "get": {
        "tags": [
          "portal"
        ],
        "operationId": "GetPortalBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/portal/{clubSlug}/bookings": {
      "post": {
        "tags": [
          "portal"
        ],
        "operationId": "CreatePortalBooking",
        "parameters": [
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalBookingCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/portal/{clubSlug}/bookings/{id}/release": {
      "post": {
        "tags": [
          "portal"
        ],
        "operationId": "ReleasePortalBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/portal/{clubSlug}/bookings/{id}/settle": {
      "post": {
        "tags": [
          "portal"
        ],
        "operationId": "SettlePortalBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalSettleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingCreatedResponse": {
        "required": [
          "id",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "BookingRequest": {
        "required": [
          "courtId",
          "date",
          "startMinute",
          "durationMinutes",
          "customerName",
          "customerPhone"
        ],
        "type": "object",
        "properties": {
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "BookingSnapshot": {
        "required": [
          "id",
          "courtId",
          "courtName",
          "sport",
          "date",
          "startMinute",
          "durationMinutes",
          "price",
          "paidAmount",
          "status",
          "paymentMode",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "courtName": {
            "type": "string"
          },
          "sport": {
            "$ref": "#/components/schemas/Sport"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "paidAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "paymentMode": {
            "$ref": "#/components/schemas/PaymentMode"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "BookingStatus": {
        "enum": [
          "confirmed",
          "cancelled",
          "pendingPayment",
          "expired"
        ],
        "type": "string"
      },
...
      "PaymentApplyOutcome": {
        "enum": [
          "confirmed",
          "rejected",
          "alreadyProcessed",
          "orphaned",
          "unknownBooking"
        ],
        "type": [
          "null",
          "string"
        ]
      },
...
      "PaymentMode": {
        "enum": [
          "club",
          "onlineFull",
          "onlineDeposit"
        ],
        "type": "string"
      },
...
      "PortalAvailability": {
        "required": [
          "currency",
          "days"
        ],
        "type": "object",
        "properties": {
          "currency": {
            "type": "string"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalDay"
            }
          }
        }
      },
      "PortalBookingCreatedResponse": {
        "required": [
          "id",
          "price",
          "chargeAmount",
          "status",
          "expiresAt",
          "checkoutUrl",
          "token"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "chargeAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "checkoutUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "token": {
            "type": "string"
          }
        }
      },
      "PortalBookingRequest": {
        "required": [
          "courtId",
          "date",
          "startMinute",
          "durationMinutes",
          "customerName",
          "customerPhone",
          "customerEmail",
          "paymentMode",
          "returnUrl"
        ],
        "type": "object",
        "properties": {
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "customerEmail": {
            "type": [
              "null",
              "string"
            ]
          },
          "paymentMode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PaymentMode"
              }
            ]
          },
          "returnUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PortalCatalogResponse": {
        "required": [
          "club",
          "sports",
          "onlinePayments"
        ],
        "type": "object",
        "properties": {
          "club": {
            "$ref": "#/components/schemas/PortalClub"
          },
          "sports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalSport"
            }
          },
          "onlinePayments": {
            "type": "boolean"
          }
        }
      },
      "PortalClub": {
        "required": [
          "name",
          "venue",
          "currency",
          "depositPercent"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "venue": {
            "type": [
              "null",
              "string"
            ]
          },
          "currency": {
            "type": "string"
          },
          "depositPercent": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PortalCourt": {
        "required": [
          "id",
          "name",
          "detail",
          "isCovered",
          "durations"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "isCovered": {
            "type": "boolean"
          },
          "durations": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "PortalDay": {
        "required": [
          "date",
          "courts"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "courts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalDayCourt"
            }
          }
        }
      },
      "PortalDayCourt": {
        "required": [
          "courtId",
          "slots"
        ],
        "type": "object",
        "properties": {
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalSlot"
            }
          }
        }
      },
      "PortalSettleResponse": {
        "required": [
          "outcome"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PaymentApplyOutcome"
              }
            ]
          }
        }
      },
      "PortalSlot": {
        "required": [
          "startMinute",
          "duration",
          "price"
        ],
        "type": "object",
        "properties": {
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PortalSport": {
        "required": [
          "sport",
          "courts"
        ],
        "type": "object",
        "properties": {
          "sport": {
            "$ref": "#/components/schemas/Sport"
          },
          "courts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalCourt"
            }
          }
        }
      },
...
      "Sport": {
        "enum": [
          "padel",
          "football"
        ],
        "type": "string"
      },
...
    }
  },
...
}